requirements.txt
streamlit>=1.30.0
yt-dlp>=2023.11.16
openai>=1.10.0
ffmpeg-python>=0.2.0
edge-tts>=6.1.9
python-dotenv>=1.0.0
pydub>=0.25.1
watchdog>=3.0.0
requests>=2.31.0
dashscope>=1.14.0
